INF-159 harden dashboard analytics evidence boundary
Expose explicit needs_data geofence evidence and move FAHP analysis builders behind a service boundary so dashboard aggregation no longer depends on controller code.
