+++ tests/summaryDashboardAnalyticsSeam.test.js
@@ -59,7 +59,7 @@ jest.unstable_mockModule('../src/utils/fuzzyAhpEngine.js', () => ({
   default: {}
 }));
 
-jest.unstable_mockModule('../src/controllers/analysis.controller.js', () => ({
+jest.unstable_mockModule('../src/services/fuzzyAhpAnalysis.service.js', () => ({
   buildDisciplineAnalysis: mockBuildDisciplineAnalysis,
   buildWfaAnalysis: mockBuildWfaAnalysis,
   buildSmartAcAnalysis: mockBuildSmartAcAnalysis
@@ -264,6 +264,8 @@ describe('summary dashboard analytics seam', () => {
         },
         geofence_evidence_context: {
           status: 'available',
+          needs_data: false,
+          reason: null,
           authority: 'context_only',
           final_attendance_authority: 'attendance_records',
           window: { from: '2026-04-01', to: '2026-04-03' },
